+++ pl_engine/pl_server/src/main/java/com/cubrid/jsp/SpSecurityManager.java
@@ -32,7 +32,7 @@
 package com.cubrid.jsp;
 
 import com.cubrid.jsp.classloader.CatalogClassLoader;
-import com.cubrid.jsp.classloader.ContextClassLoader;
+import com.cubrid.jsp.classloader.FileClassLoaderDynamic;
 import java.io.FileDescriptor;
 import java.net.InetAddress;
 import java.security.Permission;
@@ -76,7 +76,7 @@ public class SpSecurityManager extends SecurityManager {
             ClassLoader loader = cls.getClassLoader();
 
             if (loader != null
-                    && ((loader instanceof ContextClassLoader)
+                    && ((loader instanceof FileClassLoaderDynamic)
                             || (loader instanceof CatalogClassLoader))) {
                 throw new SecurityException(
                         "Native library loading is not allowed. Please load your class using 'loadjava' with the '-jni' option.");
